Bruker infotrygdutbetalinger når vi bygger snapshot

+++ src/main/kotlin/no/nav/helse/vedtaksperiode/VedtaksperiodeDto.kt
@@ -8,5 +8,5 @@ data class VedtaksperiodeDto(
     val etternavn: String,
     val arbeidsgiverRef: Long,
     val speilSnapshotRef: Long,
-    val infotrygdutbetalingerRef: Long?
+    val infotrygdutbetalingerRef: Int?
 )
